+++ forge/src/main/java/com/lx862/jcm/entrypoint/MainForge.java
@@ -4,6 +4,7 @@ import com.lx862.jcm.mod.Constants;
 import com.lx862.jcm.mod.JCM;
 import com.lx862.jcm.mod.JCMClient;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.common.Mod;
 
@@ -12,5 +13,6 @@ public class MainForge {
     public MainForge() {
         JCM.initialize();
         DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> JCMClient::initializeClient);
+        MinecraftForge.EVENT_BUS.register(new MigrateMapping());
     }
 }