+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 **Changes:**  
 
 **Fixes:**
+- All blocks will now be preserved when upgrading from JCM v1 to JCM v2 on Minecraft Forge.
 - Fix Exit sign (Even) not migrated from JCM v1
 - Fix LRT Inter Car Barriers see-through when paired with some optimization mods.